Rename Assignment3.R to CarDesign.R
a. We try to predict hipcenter with available information. Perform the best subset selection
and present the selected models with their BIC and adjusted R2in the following table.
b. Based on BIC values, which model is the best? Comment.
c. Based on adjusted R2, which model is the best? Comment.
---------------------------------------------------------------------------------------------------------------------
a. Split the data set into a training set (10%) and a test set (90%) with set.seed(1251). (Usu-
ally we do not split data like this, however, the purpose of this exercise is to show you how
regularization methods work well when training sample size is relatively small compared to the
number of variables in an extreme example. )
b. Fit a linear model using least squares on the training set, and report the test error obtained.
2
c. Fit a ridge regression model on the training set, with λ chosen by cross-validation and set.seed(19).
Report the test error obtained and chosen λ.
d. Fit a Lasso model on the training set, with λ chosen by cross-validation and set.seed(19).
Report the test error obtained, along with chosen λ and the number of zero coefficient esti-
mates.
e. Comment on the results obtained. Is there much difference among the test errors resulting
from these approaches? Which one is the best?
-------------------------------------------------------------------------------------------------------------------------
a. Use the poly() function to fit a cubic polynomial regression to predict nox using dis. Plot
the polynomial fits and 95% confidence band.
b. Plot the polynomial fits for a range of different polynomial degrees (say, from 1 to 5), and
report the associated residual sum of squares (RSS) and R2. Comment.
c. Now fit a regression spline with bs() with the knots at 4, 6, 8, and plot the resulting fit and
report the resulting RSS and R2. Why do we put knots at 4, 6, 8?
d. Compare the adjusted R2of the models in b. and c.. Which model is the best?