Rename HW#4.R to OjPrediction.R
a. Fit a tree to the training data, with Purchase as the response and the other variables as
predictors. Plot the tree. Which variables are used in the tree? How many terminal nodes
does the tree have?
b. Predict the response on the test data, and produce a confusion matrix comparing the test labels
to the predicted test labels. What is the test error rate?
c. Apply the cv.tree() function to the training set in order to determine the optimal tree size
with set.seed(984). Produce a pruned tree corresponding to the optimal tree size. Create a
plot of the tree, and interpret the tree. Interpret each terminal node.
d. Predict the response on the test data with the optimal tree size, and produce a confusion matrix.
What is the test error rate?
e. Compare the test error rates between the pruned and unpruned trees. Which is better?
f. Compare the ROC curves between the pruned and unpruned trees. Which is better?
--------------------------------------------------------------------------------------------------------------------------
a. Fit a regression tree to the training set. Plot the tree. What are the variables used? What test
MSE do you obtain?
b. Use cross-validation in order to determine the optimal level of tree complexity with set.seed(47).
Does pruning the tree improve the test MSE?
4
c. Use the bagging approach in order to analyze this data with random seed 47. What test MSE do
you obtain? Use the importance() function to determine which variables are most important.
d. Use random forests to analyze this data with random seed 47 and m = 4. What test MSE do
you obtain? Use the importance() function to determine which variables are most important.
----------------------------------------------------------------------------------------------------------------------------------------------